factored out dependency installer
